Split rendering code into one file per component
Starting to clean up the code for bringing it into the master branch.
