+++ src/rendering/component/mod.rs
@@ -1,0 +1,15 @@
+pub mod blank_space;
+pub mod current_comparison;
+pub mod current_pace;
+pub mod delta;
+pub mod detailed_timer;
+pub mod graph;
+pub mod possible_time_save;
+pub mod previous_segment;
+pub mod separator;
+pub mod splits;
+pub mod sum_of_best;
+pub mod text;
+pub mod timer;
+pub mod title;
+pub mod total_playtime;
